pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "thordata-proxy-examples"
version = "2.0.0"
description = "Complete examples for using Thordata's proxy network with the official Python SDK"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Thordata Developer Team", email = "support@thordata.com"}
]
keywords = [
    "proxy",
    "residential proxy",
    "datacenter proxy",
    "mobile proxy",
    "web scraping",
    "thordata",
    "examples",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: Proxy Servers",
]

dependencies = [
    "thordata-sdk>=1.8.4",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=25.11.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
